refactor: Separates ReXNet from ResNet
