+++ holocron/models/__init__.py
@@ -1,6 +1,5 @@
-from .res2net import *
-from .resnets import *
 from .resnet import *
+from .rexnet import *
 from .darknet import *
 from .detection import *
 from .segmentation import *
